backend/requirements.txt
Django>=5.0
psycopg[binary]>=3.2
Pillow
djangorestframework
django-filter
drf-spectacular
django-htmx
python-slugify
django-cors-headers
redis
celery
python-decouple
